+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased] - ReleaseDate
  - The log view now displays lines from both stdout and stderr
  - Add an AppLogs view to display the logs from the application using tui-logger. Bound to `L` by default.
+ - Remove auto-refresh behaviour. Use `R` to refresh a view now.
 
 ## [0.1.0] - 2018-08-28
  - Initial release